DATN

+++ K22CNT3_NVD_2210900016_DATN/K22CNT3_NVD_2210900016_DATN/Controllers/KhachHangsController.cs
@@ -8,7 +8,7 @@ using System.Web;
 using System.Web.Mvc;
 using K22CNT3_NVD_2210900016_DATN.Models;
 
-namespace K22CNT3_NVD_2210900016_DATN.Views
+namespace K22CNT3_NVD_2210900016_DATN.Controllers
 {
     public class KhachHangsController : Controller
     {
@@ -17,22 +17,19 @@ namespace K22CNT3_NVD_2210900016_DATN.Views
         // GET: KhachHangs
         public ActionResult Index()
         {
-            var list = db.KhachHangs.ToList();
-            return View(list);
+            return View(db.KhachHangs.ToList());
         }
 
         // GET: KhachHangs/Details/5
         public ActionResult Details(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             KhachHang khachHang = db.KhachHangs.Find(id);
             if (khachHang == null)
-            {
                 return HttpNotFound();
-            }
+
             return View(khachHang);
         }
 
@@ -43,11 +40,9 @@ namespace K22CNT3_NVD_2210900016_DATN.Views
         }
 
         // POST: KhachHangs/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "ID_KhachHang,TenKhach,DienThoai,Email,DiaChi")] KhachHang khachHang)
+        public ActionResult Create(KhachHang khachHang)
         {
             if (ModelState.IsValid)
             {
@@ -55,7 +50,6 @@ namespace K22CNT3_NVD_2210900016_DATN.Views
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-
             return View(khachHang);
         }
 
@@ -63,23 +57,19 @@ namespace K22CNT3_NVD_2210900016_DATN.Views
         public ActionResult Edit(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             KhachHang khachHang = db.KhachHangs.Find(id);
             if (khachHang == null)
-            {
                 return HttpNotFound();
-            }
+
             return View(khachHang);
         }
 
         // POST: KhachHangs/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "ID_KhachHang,TenKhach,DienThoai,Email,DiaChi")] KhachHang khachHang)
+        public ActionResult Edit(KhachHang khachHang)
         {
             if (ModelState.IsValid)
             {
@@ -94,14 +84,12 @@ namespace K22CNT3_NVD_2210900016_DATN.Views
         public ActionResult Delete(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             KhachHang khachHang = db.KhachHangs.Find(id);
             if (khachHang == null)
-            {
                 return HttpNotFound();
-            }
+
             return View(khachHang);
         }
 
@@ -111,17 +99,22 @@ namespace K22CNT3_NVD_2210900016_DATN.Views
         public ActionResult DeleteConfirmed(int id)
         {
             KhachHang khachHang = db.KhachHangs.Find(id);
+
+            if (khachHang == null)
+                return HttpNotFound();
+
             db.KhachHangs.Remove(khachHang);
             db.SaveChanges();
+
             return RedirectToAction("Index");
         }
+
 
         protected override void Dispose(bool disposing)
         {
             if (disposing)
-            {
                 db.Dispose();
-            }
+
             base.Dispose(disposing);
         }
     }
